rename old files, add retry for fetching fields
