+++ packages/compass-import-export/src/components/old-export/error-box/error-box.jsx
@@ -5,7 +5,7 @@ import { palette } from '@mongodb-js/compass-components';
 import ANSIConverter from 'ansi-to-html';
 
 import styles from './error-box.module.less';
-import createStyler from '../../utils/styler';
+import createStyler from '../../../utils/styler';
 const style = createStyler(styles, 'error-box');
 
 /**
