+++ packages/compass-import-export/src/export-plugin.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { withPreferences } from 'compass-preferences-model';
 
-import ExportModal from './components/export-modal';
-import exportStore from './stores/export-store';
-import { store as newExportStore } from './stores/new-export-store';
-import { ExportModal as NewExportModal } from './components/new-export-modal';
+import OldExportModal from './components/old-export/export-modal';
+import exportStore from './stores/old-export-store';
+import { store as newExportStore } from './stores/export-store';
+import { ExportModal } from './components/export-modal';
 import ExportInProgressModal from './components/export-in-progress-modal';
 
 function ExportPlugin({ useNewExport }: { useNewExport: boolean }) {
@@ -13,7 +13,7 @@ function ExportPlugin({ useNewExport }: { useNewExport: boolean }) {
   if (useNewExport) {
     return (
       <Provider store={newExportStore}>
-        <NewExportModal />
+        <ExportModal />
         <ExportInProgressModal />
       </Provider>
     );
@@ -21,7 +21,7 @@ function ExportPlugin({ useNewExport }: { useNewExport: boolean }) {
 
   return (
     <Provider store={exportStore}>
-      <ExportModal />
+      <OldExportModal />
     </Provider>
   );
 }
