+++ packages/compass-import-export/src/index.ts
@@ -1,8 +1,9 @@
 import Plugin from './plugin';
 import ImportPlugin from './import-plugin';
 import ExportPlugin from './export-plugin';
-import exportStore from './stores/export-store';
-import { store as newExportStore } from './stores/new-export-store';
+// TODO(COMPASS-6582): Rename new store and remove old.
+import oldExportStore from './stores/old-export-store';
+import { store as newExportStore } from './stores/export-store';
 import importStore from './stores/import-store';
 
 import type AppRegistry from 'hadron-app-registry';
@@ -29,7 +30,7 @@ const EXPORT_ROLE = {
  **/
 function activate(appRegistry: AppRegistry): void {
   appRegistry.registerRole('Global.Modal', EXPORT_ROLE);
-  appRegistry.registerStore('ExportModal.Store', exportStore);
+  appRegistry.registerStore('ExportModal.Store', oldExportStore);
   appRegistry.registerStore('NewExportModal.Store', newExportStore);
   appRegistry.registerRole('Global.Modal', IMPORT_ROLE);
   appRegistry.registerStore('ImportModal.Store', importStore);
