+++ packages/compass-import-export/src/modules/index.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 
 import { ns, dataService, globalAppRegistry } from './compass';
 
-import exportData from './export';
+import exportData from './old-export';
 import importData from './import';
 
 const rootReducer = combineReducers({
