+++ packages/compass-import-export/src/modules/old-export.spec.js
@@ -7,8 +7,8 @@ import PROCESS_STATUS from '../constants/process-status';
 import EXPORT_STEP from '../constants/export-step';
 import AppRegistry from 'hadron-app-registry';
 import FILE_TYPES from '../constants/file-types';
-import reducer, * as actions from './export';
-import store from '../stores/export-store';
+import reducer, * as actions from './old-export';
+import store from '../stores/old-export-store';
 import { connect } from 'mongodb-data-service';
 import { promisify } from 'util';
 import { once } from 'events';
