+++ packages/compass-import-export/src/utils/query-has-projection.ts
@@ -1,0 +1,5 @@
+import type { ExportQuery } from '../export/export-types';
+
+export function queryHasProjection(query: ExportQuery): boolean {
+  return Object.keys(query?.projection || {}).length > 0;
+}
